Changed approach, read desc
Since we live in an asynchronous world, having a large Vec<tile> is detrimental because it is impossible to control the state of the world.

An attempt was made to introduce the concept of Chunk and store these chunks, since the world has boundaries (unlike procedural generation in Minecraft), this would be the only reasonable solution.

feat(World, chunk): World contains Hashmap<Chunkid, Chunk>, chunk just Vec8<Tile>